+++ apiserver/service/src/main/java/com/zipline/service/contract/ContractService.java
@@ -4,10 +4,10 @@ import java.util.List;
 
 import org.springframework.web.multipart.MultipartFile;
 
-import com.zipline.dto.PageRequestDTO;
-import com.zipline.dto.contract.ContractListResponseDTO;
-import com.zipline.dto.contract.ContractRequestDTO;
-import com.zipline.dto.contract.ContractResponseDTO;
+import com.zipline.global.request.PageRequestDTO;
+import com.zipline.service.contract.dto.response.ContractListResponseDTO;
+import com.zipline.service.contract.dto.request.ContractRequestDTO;
+import com.zipline.service.contract.dto.response.ContractResponseDTO;
 
 public interface ContractService {
