+++ apiserver/service/src/main/java/com/zipline/service/contract/dto/response/ContractResponseDTO.java
@@ -1,4 +1,4 @@
-package com.zipline.dto.contract;
+package com.zipline.service.contract.dto.response;
 
 import java.time.LocalDate;
 import java.util.List;
